# packages/typescript/api-types/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Roborean API",
    "version": "0.4.0"
  },
  "paths": {
...
    "/v1/projects/{project_id}/templates/{template_id}/content": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Template Content",
        "description": "Fetch one template file for a stored project.\n\nArgs:\n    project_id: Identifier of the project to load.\n    template_id: Template identifier from the project table.\n    state: Shared repositories and run service.\n    _principal: Resolved caller identity (auth stub).\n\nReturns:\n    Template file bytes and optional UTF-8 text body.",
        "operationId": "get_template_content_v1_projects__project_id__templates__template_id__content_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateContentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "projects"
        ],
        "summary": "Put Template Content",
        "description": "Replace one template file for a stored project.\n\nArgs:\n    project_id: Identifier of the project to update.\n    template_id: Template identifier from the project table.\n    body: Template bytes or UTF-8 text to persist.\n    state: Shared repositories and run service.\n    _principal: Resolved caller identity (auth stub).\n\nReturns:\n    Stored template file metadata.",
        "operationId": "put_template_content_v1_projects__project_id__templates__template_id__content_put",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateContentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateContentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Template Content",
        "description": "Delete one template file from project storage.\n\nArgs:\n    project_id: Identifier of the project to update.\n    template_id: Template identifier from the project table.\n    state: Shared repositories and run service.\n    _principal: Resolved caller identity (auth stub).\n\nReturns:\n    Empty response with HTTP 204 status.",
        "operationId": "delete_template_content_v1_projects__project_id__templates__template_id__content_delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/template-library": {
      "get": {
        "tags": [
          "template-library"
        ],
        "summary": "List Template Library",
        "description": "List catalog entries with optional filters.\n\nArgs:\n    kind: Optional entry kind filter.\n    document_type: Optional document type filter.\n    tag: Optional tag filter.\n    _principal: Resolved caller identity (auth stub).\n\nReturns:\n    Matching catalog summary rows.",
        "operationId": "list_template_library_v1_template_library_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "documentType",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Documenttype"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateLibraryEntry"
                  },
                  "title": "Response List Template Library V1 Template Library Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/template-library/{entry_id}": {
      "get": {
        "tags": [
          "template-library"
        ],
        "summary": "Get Template Library Entry",
        "description": "Fetch one catalog entry with its payload body.\n\nArgs:\n    entry_id: Catalog entry identifier.\n    _principal: Resolved caller identity (auth stub).\n\nReturns:\n    Entry summary plus manifest, project, or recipe payload.\n\nRaises:\n    ApiError: When the entry id is unknown.",
        "operationId": "get_template_library_entry_v1_template_library__entry_id__get",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateLibraryDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/template-library/{entry_id}/content": {
      "get": {
        "tags": [
          "template-library"
        ],
        "summary": "Get Template Library Content",
        "description": "Return template bytes for a document catalog entry.\n\nArgs:\n    entry_id: Document catalog entry identifier.\n    _principal: Resolved caller identity (auth stub).\n\nReturns:\n    Base64-encoded template bytes plus a UTF-8 hint when applicable.\n\nRaises:\n    ApiError: When the entry is missing or not a document template.",
        "operationId": "get_template_library_content_v1_template_library__entry_id__content_get",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateContentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RequiredBitTypeSummary": {
        "properties": {
          "typeId": {
            "type": "string",
            "title": "Typeid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "typeId",
          "name"
        ],
        "title": "RequiredBitTypeSummary",
        "description": "Required bit type row for recipe catalog entries.\n\nAttributes:\n    type_id: Bit type identifier (``typeId`` alias).\n    name: Human-readable bit type name."
      },
...
      "TemplateContentResponse": {
        "properties": {
          "templateId": {
            "type": "string",
            "title": "Templateid"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "contentBase64": {
            "type": "string",
            "title": "Contentbase64"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "templateId",
          "path",
          "contentBase64"
        ],
        "title": "TemplateContentResponse",
        "description": "Template file payload returned by the API.\n\nAttributes:\n    template_id: Template identifier from the project table\n        (``templateId`` alias).\n    path: Relative package path for the template file.\n    content_base64: Base64-encoded template bytes\n        (``contentBase64`` alias).\n    text: UTF-8 decoded body when the file is textual."
      },
      "TemplateContentUpdate": {
        "properties": {
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "contentBase64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contentbase64"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TemplateContentUpdate",
        "description": "Replace one template file.\n\nAttributes:\n    text: UTF-8 template body for text-like templates.\n    content_base64: Base64-encoded bytes for binary templates\n        (``contentBase64`` alias)."
      },
      "TemplateLibraryDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "document",
              "starter",
              "recipe"
            ],
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "documentType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documenttype"
          },
          "driver": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver"
          },
          "irFamily": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Irfamily"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "templateVersion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Templateversion"
          },
          "requiredInputs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requiredinputs"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "mediaType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mediatype"
          },
          "variableCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variablecount"
          },
          "bitCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bitcount"
          },
          "documentCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documentcount"
          },
          "requiredBitTypes": {
            "items": {
              "$ref": "#/components/schemas/RequiredBitTypeSummary"
            },
            "type": "array",
            "title": "Requiredbittypes"
          },
          "manifest": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifest"
          },
          "project": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          },
          "recipe": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipe"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "title"
        ],
        "title": "TemplateLibraryDetail",
        "description": "Full catalog entry with optional payload bodies.\n\nAttributes:\n    manifest: Parsed template manifest for document entries.\n    project: Starter project document for starter entries.\n    recipe: Recipe document for recipe entries."
      },
      "TemplateLibraryEntry": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "document",
              "starter",
              "recipe"
            ],
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "documentType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documenttype"
          },
          "driver": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver"
          },
          "irFamily": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Irfamily"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "templateVersion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Templateversion"
          },
          "requiredInputs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requiredinputs"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "mediaType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mediatype"
          },
          "variableCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variablecount"
          },
          "bitCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bitcount"
          },
          "documentCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documentcount"
          },
          "requiredBitTypes": {
            "items": {
              "$ref": "#/components/schemas/RequiredBitTypeSummary"
            },
            "type": "array",
            "title": "Requiredbittypes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "title"
        ],
        "title": "TemplateLibraryEntry",
        "description": "Summary row for one catalog entry.\n\nAttributes:\n    id: Stable catalog entry identifier.\n    kind: Entry kind: document, starter, or recipe.\n    title: Human-readable catalog title.\n    description: Optional longer description.\n    document_type: Document type for document entries\n        (``documentType`` alias).\n    driver: Document driver id for document entries.\n    ir_family: Intermediate representation family for documents\n        (``irFamily`` alias).\n    tags: Optional discovery tags.\n    template_version: Template semver for document entries\n        (``templateVersion`` alias).\n    required_inputs: Named template inputs for document entries\n        (``requiredInputs`` alias).\n    capabilities: Declared document capabilities for document entries.\n    path: Relative template path for document entries.\n    media_type: MIME type for document template bytes\n        (``mediaType`` alias).\n    variable_count: Variable count for starters and recipes\n        (``variableCount`` alias).\n    bit_count: Bit count for starters and recipes (``bitCount`` alias).\n    document_count: Document count for starters and recipes\n        (``documentCount`` alias).\n    required_bit_types: Required bit types for recipe entries\n        (``requiredBitTypes`` alias)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}